// models/stage_group.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0'
 */
export type StageGroup = IContentItem<{
  title: Elements.TextElement;
  order: Elements.NumberElement;
  years: Elements.MultipleChoiceElement;
  stages: Elements.LinkedItemsElement<IContentItem>;
}>;